requirements.txt
certifi==2020.6.20
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.16.0
MarkupSafe==1.1.1
numpy==1.19.1
psycopg2==2.8.5
psycopg2-binary==2.8.5
python-dotenv==0.14.0
scikit-learn==0.23.1
scipy==1.5.2
six==1.15.0
sklearn==0.0
SQLAlchemy==1.3.18
threadpoolctl==2.1.0
Werkzeug==1.0.1
